% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelNMoE.R
\docType{class}
\name{ModelNMoE-class}
\alias{ModelNMoE-class}
\alias{ModelNMoE}
\title{A Reference Class which represents a fitted NMoE model.}
\description{
ModelNMoE represents a \link[=ModelNMoE]{NMoE} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamNMoE} object. It contains the estimated values of the parameters.}

\item{\code{stat}}{A \link{StatNMoE} object. It contains all the statistics associated to the NMoE model.}
}}


\seealso{
\link{ParamNMoE}, \link{StatNMoE}
}
