% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSNMoE.R
\docType{class}
\name{ModelSNMoE-class}
\alias{ModelSNMoE-class}
\alias{ModelSNMoE}
\title{A Reference Class which represents a fitted SNMoE model.}
\description{
ModelSNMoE represents a \link[=ModelSNMoE]{SNMoE} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamSNMoE} object. It contains the estimated values of the parameters.}

\item{\code{stat}}{A \link{StatSNMoE} object. It contains all the statistics associated to the SNMoE model.}
}}


\seealso{
\link{ParamSNMoE}, \link{StatSNMoE}
}
