% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTMoE.R
\docType{class}
\name{ModelTMoE-class}
\alias{ModelTMoE-class}
\alias{ModelTMoE}
\title{A Reference Class which represents a fitted TMoE model.}
\description{
ModelTMoE represents a \link[=ModelTMoE]{TMoE} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamTMoE} object. It contains the estimated values of the parameters.}

\item{\code{stat}}{A \link{StatTMoE} object. It contains all the statistics associated to the TMoE model.}
}}


\seealso{
\link{ParamTMoE}, \link{StatTMoE}
}
