% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTMoE.R
\docType{class}
\name{ParamTMoE-class}
\alias{ParamTMoE-class}
\alias{ParamTMoE}
\title{A Reference Class which contains parameters of a TMoE model.}
\description{
ParamTMoE contains all the parameters of a TMoE model.
}
\section{Fields}{

\describe{
\item{\code{X}}{Numeric vector of length \emph{n} representing the covariates/inputs
\eqn{x_{1},\dots,x_{n}}.}

\item{\code{Y}}{Numeric vector of length \emph{n} representing the observed
response/output \eqn{y_{1},\dots,y_{n}}.}

\item{\code{n}}{Numeric. Length of the response/output vector \code{Y}.}

\item{\code{K}}{The number of mixture components.}

\item{\code{p}}{The order of the polynomial regression.}

\item{\code{q}}{The dimension of the logistic regression. For the purpose of
segmentation, it must be set to 1.}

\item{\code{df}}{degree of freedom}

\item{\code{alpha}}{is the parameter vector of the logistic model with \eqn{alpha_K} being the null vector.}

\item{\code{beta}}{is the vector of regression coefficients of component k,
the updates for each of the expert component parameters consist in analytically solving a weighted
Gaussian linear regression problem.}

\item{\code{sigma2}}{The variances for the \emph{K} mixture component.}

\item{\code{nu}}{degrees of freedom}
}}

\section{Methods}{

\describe{
\item{\code{initParam(try_EM, segmental = FALSE)}}{Method to initialize parameters \code{alpha}, \code{beta} and
\code{sigma2}.}
}}

