% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simulatedstructureddata.R
\docType{data}
\name{simulatedstructureddata}
\alias{simulatedstructureddata}
\title{Toy sample}
\format{An object of class \code{data.frame} with 500 rows and 2 columns.}
\usage{
simulatedstructureddata
}
\description{
Toy sample
}
\keyword{datasets}
