% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatNMoE.R
\docType{class}
\name{StatNMoE-class}
\alias{StatNMoE-class}
\alias{StatNMoE}
\title{A Reference Class which contains statistics of a NMoE model.}
\description{
StatNoE contains all the parameters of a \link[=ParamNMoE]{NMoE} model.
}
\section{Fields}{

\describe{
\item{\code{piik}}{Matrix of size \eqn{(n, K)} representing the probabilities
\eqn{P(zi = k; W) = P(z_{ik} = 1; W)}{P(zi = k; W) = P(z_ik = 1; W)} of the
latent variable \eqn{zi,\ i = 1,\dots,m}{zi, i = 1,\dots,n}.}

\item{\code{z_ik}}{Hard segmentation logical matrix of dimension \eqn{(n, K)}
obtained by the Maximum a posteriori (MAP) rule: \eqn{z_{ik} = 1 \
  \textrm{if} \ z_{ik} = \textrm{arg} \ \textrm{max}_{k} \ P(z_i = k | Y, W,
  \beta);\ 0 \ \textrm{otherwise}}{z_ik = 1 if z_ik = arg max_k P(z_i = k |
  Y, W, \beta); 0 otherwise}, \eqn{k = 1,\dots,K}.}

\item{\code{klas}}{Column matrix of the labels issued from \code{z_ik}. Its elements are
\eqn{klas(i) = k}, \eqn{k = 1,\dots,K}.}

\item{\code{Wik}}{Matrix of dimension \emph{(nm,K)}.}

\item{\code{Ey_k}}{Matrix of dimension \emph{(n,K)}.}

\item{\code{Ey}}{Column matrix of dimension \emph{n}.}

\item{\code{Var_yk}}{Column matrix of dimension \emph{K}.}

\item{\code{Vary}}{Column matrix of dimension \emph{n}.}

\item{\code{log\_lik}}{Numeric. Log-likelihood of the StMoE model.}

\item{\code{com_loglik}}{Numeric. Complete log-likelihood of the StMoE model.}

\item{\code{stored_loglik}}{Numeric vector. Stored values of the log-likelihood at
each EM iteration.}

\item{\code{BIC}}{Numeric. Value of the BIC (Bayesian Information Criterion)
criterion. The formula is \eqn{BIC = log\_lik - nu \times \textrm{log}(n) /
  2}{BIC = log\_lik - nu x log(n) / 2} with \emph{nu} the degree of freedom
of the StMoE model.}

\item{\code{ICL}}{Numeric. Value of the ICL (Integrated Completed Likelihood)
criterion. The formula is \eqn{ICL = com\_loglik - nu \times
  \textrm{log}(n) / 2}{ICL = com_loglik - nu x log(n) / 2} with \emph{nu} the
degree of freedom of the StMoE model.}

\item{\code{AIC}}{Numeric. Value of the AIC (Akaike Information Criterion)
criterion. The formula is \eqn{AIC = log\_lik - nu}{AIC = log\_lik - nu}.}

\item{\code{log_piik_fik}}{Matrix of size \eqn{(n, K)} giving the values of the
logarithm of the joint probability \eqn{P(Y_{i}, \ zi = k)}{P(Yi, zi = k)},
\eqn{i = 1,\dots,n}.}

\item{\code{log_sum_piik_fik}}{Column matrix of size \emph{n} giving the values of
\eqn{\sum_{k = 1}^{K} \textrm{log} P(Y_{i}, \ zi = k)}{\sum_{k = 1}^{K} log
  P(Yi, zi = k)}, \eqn{i = 1,\dots,n}.}

\item{\code{tik}}{Matrix of size \eqn{(n, K)} giving the posterior probability that
\eqn{Y_{i}}{Yi} originates from the \eqn{k}-th regression model \eqn{P(zi =
  k | Y, W, \beta)}.}
}}

\section{Methods}{

\describe{
\item{\code{MAP()}}{calcule une partition d'un echantillon par la regle du Maximum A Posteriori ?? partir des probabilites a posteriori
Entrees : post_probas , Matrice de dimensions [n x K] des probabibiltes a posteriori (matrice de la partition floue)
n : taille de l'echantillon
K : nombres de classes
klas(i) = arg   max (post_probas(i,k)) , for all i=1,...,n
1<=k<=K
= arg   max  p(zi=k|xi;theta)
1<=k<=K
= arg   max  p(zi=k;theta)p(xi|zi=k;theta)/sum{l=1}^{K}p(zi=l;theta) p(xi|zi=l;theta)
1<=k<=K
Sorties : classes : vecteur collones contenant les classe (1:K)
Z : Matrice de dimension [nxK] de la partition dure : ses elements sont zik, avec zik=1 si xi
appartient ?? la classe k (au sens du MAP) et zero sinon.}
}}

\seealso{
\link{ParamNMoE}, \link{FData}
}
