% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamStMoE.R
\docType{class}
\name{ParamStMoE-class}
\alias{ParamStMoE-class}
\alias{ParamStMoE}
\title{A Reference Class which contains parameters of a MRHLP model.}
\description{
ParamMRHLP contains all the parameters of a MRHLP model.
}
\section{Fields}{

\describe{
\item{\code{fData}}{\link{FData} object representing the sample.}

\item{\code{K}}{The number of mixture components.}

\item{\code{p}}{The order of the polynomial regression.}

\item{\code{q}}{The dimension of the logistic regression. For the purpose of
segmentation, it must be set to 1.}

\item{\code{df}}{degree of freedom}

\item{\code{alpha}}{is the parameter vector of the logistic model with \eqn{alpha_K}
being the null vector.}

\item{\code{beta}}{is the vector of regression coefficients of component k, the
updates for each of the expert component parameters consist in analytically
solving a weighted Gaussian linear regression problem.}

\item{\code{sigma2}}{The variances for the \emph{K} mixture component.}

\item{\code{lambda}}{skewness parameter}

\item{\code{delta}}{the skewness parameter lambda (by equivalence delta)}

\item{\code{nuk}}{degrees of freedom}
}}

\section{Methods}{

\describe{
\item{\code{initParam(try_EM, segmental = FALSE)}}{Method to initialize parameters \code{alpha}, \code{beta} and
\code{sigma2}.}

\item{\code{MStep(statStMoE, calcAlpha = FALSE, calcBeta = FALSE, calcSigma2 = FALSE,
  calcLambda = FALSE, calcNu = FALSE, verbose_IRLS = FALSE)}}{Method used in the EM algorithm to learn the parameters of the StMoE model
based on statistics provided by \code{statStMoE}.}
}}

\seealso{
\link{FData}
}
